🚚 move tps report logic into bot module src
